fix: Changed radius in area from decimal to int
Also changed all tests using create_area. Now using PositiveIntegerField.
Added new test bad radius to check that a negative radius is not allowed.

+++ backend/database_wrapper.py
@@ -898,7 +898,7 @@ def get_areas_in_filter(fid: int) -> List[Area]:
     return f.areas.all()[::1]
 
 
-def create_area(latitude: Decimal, longitude: Decimal, radius: Decimal, fid: int) -> int:
+def create_area(latitude: Decimal, longitude: Decimal, radius: int, fid: int) -> int:
     """
     Creates an area and adds to filter.
